day 2

+++ src/basic_cheeze/ArithmeticOperationsOnInt.java
@@ -2,7 +2,7 @@ package basic_cheeze;
 
 public class ArithmeticOperationsOnInt {
     static void main() {
-        float x = 5 , y = 6;
+        double x = 5 , y = 6;
         System.out.println(x+y);
         System.out.println(x-y);
         System.out.println(x*y);
